+++ changes.md
@@ -19,6 +19,7 @@ output:
 + `tradeoff_batch` passes out the bcode now
 + `pareto` uses bcode passed from `tradeoff_batch`, and saves optitab to batch directory. 
 + `tradeoff_batch` impact tables now come out as data.frames from both todisk TRUE and FALSE cases.  Changes to `batch_param` made to accomodate (table rather than list). From the todisk=FALSE case, the output conversion tables are served as a list that is the second element of the output list. 
++ Added `imp_sorter` function for selecting subsets of conversion maps and ordering them to correctly match names of selected impact tables. 
 
 ## Bugs
 + Removed hard-coded `input_key` ("ZA") in `tradeoff_batch` call to `tradeoff_mod`.